refactor: :recycle: created dedicated endpoint for /user/renew
